+++ NEWS.md
@@ -1,3 +1,9 @@
+# rdeck dev
+
+- Add feature editor (#75)
+- Rewrite client api (#73)
+
+
 # rdeck 0.4.0
 
 - Add layer partial updaters for shiny